# 01/spec/api.yml
openapi: 3.0.3
info:
  title: 力麗飯店 LINE OA CRM API
  description: 多渠道會員管理與行銷平台 API，整合 LINE、Facebook Messenger 和 Webchat。
  version: 0.6.0
servers:
  - url: /api/v1
    description: API v1

# ============================================================================
# 既有功能 API（根據現有代碼與 Feature 規格）
# ============================================================================
paths:
...
  /chatbot/message:
    post:
      summary: 官網聊天機器人發送訊息（含意圖分析、FAQ 引用、PMS 即時查詢）
      operationId: chatbotSendMessage
      tags: [官網聊天機器人]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [browser_key, message]
              properties:
                browser_key:
                  type: string
                  description: 瀏覽器唯一識別碼（Web Key）
                message:
                  type: string
                  description: 訪客訊息內容
                hotel_id:
                  type: integer
                  description: 飯店 ID（多租戶場景）
      responses:
        '200':
          description: AI 回覆成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                  intent_state:
                    type: string
                    enum: [detecting, confirmed, none]
                  reply:
                    type: string
                    description: AI 回覆文字
                  reply_type:
                    type: string
                    enum: [text, room_cards, member_form, booking_confirm]
                    description: 回覆類型，前端依此渲染對應 UI
                  room_cards:
                    type: array
                    description: 房型推薦卡片（reply_type = room_cards 時）
                    items:
                      $ref: '#/components/schemas/RoomCard'
                  missing_fields:
                    type: array
                    description: 尚缺的訂房資訊欄位
                    items:
                      type: string
                  turn_count:
                    type: integer
                  booking_context:
                    type: object
                    description: 目前已解析出的訂房上下文
                    properties:
                      checkin_date:
                        type: string
                        format: date
                        nullable: true
                      checkout_date:
                        type: string
                        format: date
                        nullable: true
                      adults:
                        type: integer
                        nullable: true
                      rooms:
                        type: array
                        description: 已解析的幾間幾人房請求（room_plan_requests）
                        nullable: true
                        items:
                          type: object
                          properties:
                            room_count:
                              type: integer
                              example: 1
                            adults_per_room:
                              type: integer
                              example: 2
  /chatbot/rooms:
    get:
      summary: 即時查詢符合條件的空房型（呼叫 PMS API）
      operationId: chatbotGetRooms
      tags: [官網聊天機器人]
      parameters:
        - name: browser_key
          in: query
          required: true
          schema:
            type: string
        - name: checkin_date
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2026-03-07"
        - name: checkout_date
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2026-03-09"
        - name: adults
          in: query
          required: true
          schema:
            type: integer
        - name: children
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 房型清單
          content:
            application/json:
              schema:
                type: object
                properties:
                  source:
                    type: string
                    enum: [pms, faq_static]
                    description: 資料來源（pms=即時 / faq_static=靜態降級；靜默切換，前台不顯示來源）
                  rooms:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoomCard'
                  auto_split_options:
                    type: array
                    nullable: true
                    description: 混搭推薦方案（_auto_split_options，單一房型不足時由後端自動計算）
                    items:
                      type: object
                      properties:
                        label:
                          type: string
                          example: "建議 2 間雙人房"
                        rooms:
                          type: array
                          items:
                            $ref: '#/components/schemas/SelectedRoomItem'
        # <!-- ?1[缺漏] 查詢空房型.feature ?1：無任何房型時是否提供「留資料待客服聯繫」入口？目前回傳 rooms=[] 並由 /chatbot/message 引導重填，尚未決定是否開啟備用收單路徑 -->
        '422':
          description: booking_context 不完整（missing checkin_date / checkout_date / room_plan_requests）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatbotError'
  /chatbot/confirm-room:
    post:
      summary: 訪客確認房型間數選取，系統更新 Session 並回傳會員資料表單欄位定義
      operationId: chatbotConfirmRoom
      tags: [官網聊天機器人]
      description: |
        房型卡片的間數選取為純前端狀態。
        訪客按「確認選擇」後，前端呼叫此 endpoint 寫入 Session，
        並取得 member_form 的欄位結構定義（跳過 AI 對話，直接進入會員資料搜集步驟）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [browser_key, rooms]
              properties:
                browser_key:
                  type: string
                rooms:
                  type: array
                  description: 已選擇的房型與間數清單，僅提交 room_count > 0 的房型
                  items:
                    $ref: '#/components/schemas/SelectedRoomItem'
      responses:
        '200':
          description: Session 更新成功，回傳會員表單欄位定義
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply_type:
                    type: string
                    enum: [member_form]
                  session_id:
                    type: string
                  selected_room_type:
                    type: string
                    description: 向下相容欄位（selected_rooms[0].room_type_code）
                  selected_room_count:
                    type: integer
                    description: 向下相容欄位（selected_rooms[0].room_count）
                  selected_rooms:
                    type: array
                    description: 完整多房型記錄
                    items:
                      $ref: '#/components/schemas/SelectedRoomItem'
                  member_form:
                    $ref: '#/components/schemas/MemberFormDefinition'
        '422':
          description: 未選任何房型（rooms 為空陣列）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatbotError'
              example:
                error_code: NO_ROOMS_SELECTED
                message: 請至少選擇一個房型與間數
  /chatbot/session/reset:
    post:
      summary: 重置官網聊天機器人的當前 Session
      operationId: chatbotResetSession
      tags: [官網聊天機器人]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [browser_key]
              properties:
                browser_key:
                  type: string
      responses:
        '200':
          description: Session 已重置
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  session_id:
                    type: string
  /chatbot/booking-save:
    post:
      summary: 確認並儲存官網訪客的訂房資料
      operationId: chatbotSaveBooking
      tags: [官網聊天機器人]
      description: |
        以現行聊天機器人流程為準：
        前端在訪客完成房型選擇與會員資料填寫後，直接呼叫此 endpoint。
        後端會驗證 Session 中的訂房上下文，並寫入 members / chatbot_sessions / booking_records；
        若資料庫不可用，則降級寫入本地 reservation 檔案。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [browser_key, name, phone, email]
              properties:
                browser_key:
                  type: string
                name:
                  type: string
                phone:
                  type: string
                  pattern: '^\\d{10}$'
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: 訂房資料保存成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reservation_id:
                    type: string
                  saved:
                    type: object
                    properties:
                      reservation_id:
                        type: string
                      browser_key:
                        type: string
                      session_id:
                        type: string
                      checkin_date:
                        type: string
                        format: date
                      checkout_date:
                        type: string
                        format: date
                      adults:
                        type: integer
                      children:
                        type: integer
                      selected_rooms:
                        type: array
                        items:
                          $ref: '#/components/schemas/SelectedRoomItem'
                      room_type_code:
                        type: string
                        nullable: true
                      room_type_name:
                        type: string
                        nullable: true
                      room_source:
                        type: string
                        enum: [pms, faq_static]
                      room_count:
                        type: integer
                      member_name:
                        type: string
                      member_phone:
                        type: string
                      member_email:
                        type: string
                      crm_member_id:
                        type: integer
                        nullable: true
                      db_saved:
                        type: boolean
                      created_at:
                        type: string
                        format: date-time
                  cart_url:
                    type: string
                    nullable: true
                    description: PMS 購物車 URL（booking-save 同時呼叫 PMS 建立購物車並回傳；測試模式 ENABLE_DB=false 時為 null）
                    example: "https://pms.example.com/cart?reservation_id=xxx"
        '422':
          description: 訂房上下文不完整或會員資料格式錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatbotError'
              examples:
                incomplete_context:
                  value:
                    error_code: INCOMPLETE_BOOKING_CONTEXT
                    message: "缺少必要欄位：checkin_date"
                invalid_phone:
                  value:
                    error_code: INVALID_PHONE
                    message: "電話格式錯誤，需為 10 位數字"
                invalid_date_range:
                  value:
                    error_code: INVALID_DATE_RANGE
                    message: "退房日期須晚於入住日期"
        '500':
          description: 資料庫寫入失敗

  # --------------------------------------------------------------------------
  # Chatbot PMS 串接管理
  # Feature: chatbot 串接 PMS
  # --------------------------------------------------------------------------
  /chatbot/pms-status:
    get:
      summary: 取得 PMS 串接啟用狀態
      operationId: getChatbotPmsStatus
      tags: [官網聊天機器人]
      responses:
        '200':
          description: PMS 狀態
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
                    description: 是否啟用 PMS 串接
                  configured:
                    type: boolean
                    description: PMS 環境變數是否已設定
                  last_synced_at:
                    type: string
                    nullable: true
                    description: 最近同步時間（格式 YYYY-MM-DD HH:MM）
    put:
      summary: 切換 PMS 串接啟用狀態
      operationId: updateChatbotPmsStatus
      tags: [官網聊天機器人]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
        '422':
          description: 缺少 enabled 欄位
  /chatbot/pms-validate-room:
    post:
      summary: 檢核房型代碼在 PMS 中是否仍有效
      operationId: validateChatbotPmsRoom
      tags: [官網聊天機器人]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [room_code]
              properties:
                room_code:
                  type: string
      responses:
        '200':
          description: 檢核結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  message:
                    type: string
                  available_codes:
                    type: array
                    items:
                      type: string
                    description: 當 valid=false 時回傳 PMS 中可用的房型代碼
        '422':
          description: 缺少 room_code

  # --------------------------------------------------------------------------
  # FAQ 後台 PMS 串接設定 (faq pms connection)
  # --------------------------------------------------------------------------
...
components:
  schemas:
...
    ChatbotError:
      type: object
      description: 聊天機器人 API 錯誤回應
      required: [error_code, message]
      properties:
        error_code:
          type: string
          enum:
            - INCOMPLETE_BOOKING_CONTEXT
            - INVALID_PHONE
            - INVALID_DATE_RANGE
            - NO_ROOMS_SELECTED
          description: 機器可讀的錯誤代碼
        message:
          type: string
          description: 人類可讀的錯誤說明（可含缺少欄位列表）
          example: "缺少必要欄位：checkin_date"
    RoomCard:
      type: object
      description: 房型推薦卡片（用於聊天機器人 UI）
      properties:
        room_type_code:
          type: string
        room_type_name:
          type: string
        price:
          type: number
          description: 房價（PMS 即時價 / FAQ 參考價）
        price_label:
          type: string
          description: 顯示標籤，如「即時房價」或「一般參考房價」
        available_count:
          type: integer
          nullable: true
          description: 剩餘間數（PMS 有值；FAQ 靜態資料時為 null）
        max_occupancy:
          type: integer
          description: 最多可入住人數
        image_url:
          type: string
          nullable: true
        features:
          type: string
          description: 房型特色文字（來自 FAQ 規則）
        source:
          type: string
          enum: [pms, faq_static]
    SelectedRoomItem:
      type: object
      description: 前端確認送出的房型與間數資料
      properties:
        room_type_code:
          type: string
        room_count:
          type: integer
          minimum: 1
        room_type_name:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
          enum: [pms, faq_static]
...
    MemberFormDefinition:
      type: object
      description: 會員資料表單的欄位定義（由 /chatbot/confirm-room 回傳，前端依此渲染表單）
      properties:
        fields:
          type: array
          items:
            type: object
            properties:
              field_name:
                type: string
                enum: [guest_name, guest_phone, guest_email]
              label:
                type: string
                example: 姓名
              is_required:
                type: boolean
              input_type:
                type: string
                enum: [text, tel, email]
              validation_pattern:
                type: string
                description: 前端格式驗證正則，如電話 ^\d{10}$
              error_message:
                type: string
                description: 格式錯誤時顯示的提示文字
        privacy_note:
          type: string
          description: 表單下方小字提示
          example: "您的個資僅用於本次訂房聯繫與 CRM 會員服務，請安心填寫。"
...